+++ src/main/java/com/sparta/twitNation/util/api/ApiResult.java
@@ -1,4 +1,4 @@
-package com.sparta.twitNation.util.aop;
+package com.sparta.twitNation.util.api;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
